added JWT

+++ resources/item.py
@@ -3,26 +3,35 @@ from flask_smorest import Blueprint, abort
 from schemas import ItemSchema, ItemUpdateSchema
 from models import ItemModel
 from db import db
+from flask_jwt_extended import jwt_required, get_jwt
 from sqlalchemy.exc import SQLAlchemyError
 
 
 blp = Blueprint("items", __name__, description="operations on items")
 
 
-@blp.route("/item/<string:item_id>")
+@blp.route("/item/<int:item_id>")
 class Item(MethodView):
     @blp.response(200, ItemSchema)
     def get(self, item_id):
         item = ItemModel.query.get_or_404(item_id)
         return item
 
+    @jwt_required(fresh=True)
+    @blp.doc(security=[{"jwt": []}])
     def delete(self, item_id):
+        jwt = get_jwt()
+        if not jwt.get("is_admin"):
+            abort(401, message="Admin privilege required.")
+
         item = ItemModel.query.get_or_404(item_id)
         db.session.delete(item)
         db.session.commit()
 
         return {"message": "Item deleted successfully."}
 
+    @jwt_required()
+    @blp.doc(security=[{"jwt": []}])
     @blp.arguments(ItemUpdateSchema)
     @blp.response(200, ItemSchema)
     def put(self, item_data, item_id):
@@ -45,6 +54,8 @@ class ItemList(MethodView):
     def get(self):
         return ItemModel.query.all()
 
+    @jwt_required()
+    @blp.doc(security=[{"jwt": []}])
     @blp.arguments(ItemSchema)
     @blp.response(201, ItemSchema)
     def post(self, item_data):
